+++ service/controller/app/v1/resource/appcatalog/resource.go
@@ -51,34 +51,6 @@ func New(config Config) (*Resource, error) {
 	return r, nil
 }
 
-func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
-	cr, err := key.ToCustomResource(obj)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = r.getCatalogForApp(ctx, cr)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	return nil
-}
-
-func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
-	cr, err := key.ToCustomResource(obj)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = r.getCatalogForApp(ctx, cr)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	return nil
-}
-
 func (*Resource) Name() string {
 	return Name
 }
