+++ service/controller/app/v1/resource/appnamespace/create.go
@@ -1,0 +1,7 @@
+package appnamespace
+
+import "context"
+
+func (*Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
+	return nil
+}
