+++ service/controller/app/v1/resource/clients/create.go
@@ -1,0 +1,23 @@
+package clients
+
+import (
+	"context"
+
+	"github.com/giantswarm/microerror"
+
+	"github.com/giantswarm/app-operator/service/controller/app/v1/key"
+)
+
+func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
+	cr, err := key.ToCustomResource(obj)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	err = r.addClientsToContext(ctx, cr)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	return nil
+}
